Business Test Cases

+++ Customers/Bookmyslot.Api.Customers.Business/CustomerBusiness.cs
@@ -5,6 +5,7 @@ using Bookmyslot.Api.Customers.Contracts.Interfaces;
 using FluentValidation.Results;
 using System.Collections.Generic;
 using System.Threading.Tasks;
+using System.Linq;
 
 namespace Bookmyslot.Api.Customers.Business
 {
@@ -23,14 +24,24 @@ namespace Bookmyslot.Api.Customers.Business
 
             if (results.IsValid)
                 return await customerRepository.CreateCustomer(customerModel);
-            
+
             else
-                return new Response<bool>() { ResultType = ResultType.ValidationError };
+                return new Response<bool>() { ResultType = ResultType.ValidationError, Messages = results.Errors.Select(a=>a.ErrorMessage).ToList() };
         }
 
         public async Task<Response<bool>> DeleteCustomer(string email)
         {
-            return await this.customerRepository.DeleteCustomer(email);
+            if (string.IsNullOrWhiteSpace(email))
+            {
+                return new Response<bool>() { ResultType = ResultType.ValidationError, Messages = new List<string>() { Constants.EmailIdNotValid } };
+            }
+
+            var customerExists = await CheckIfCustomerExists(email);
+            if (customerExists)
+            {
+                return await this.customerRepository.DeleteCustomer(email);
+            }
+            return new Response<bool>() { ResultType = ResultType.Error, Messages = new List<string>() { Constants.CustomerNotFound } };
         }
 
         public async Task<Response<IEnumerable<CustomerModel>>> GetAllCustomers()
@@ -40,13 +51,32 @@ namespace Bookmyslot.Api.Customers.Business
 
         public async Task<Response<CustomerModel>> GetCustomer(string email)
         {
+            if (string.IsNullOrWhiteSpace(email))
+            {
+                return new Response<CustomerModel>() { ResultType = ResultType.ValidationError, Messages = new List<string>() { Constants.EmailIdNotValid } };
+            }
 
             return await customerRepository.GetCustomer(email);
         }
 
         public async Task<Response<bool>> UpdateCustomer(CustomerModel customerModel)
         {
-            return await this.customerRepository.UpdateCustomer(customerModel);
+            var customerExists = await CheckIfCustomerExists(customerModel.Email);
+            if (customerExists)
+            {
+                return await this.customerRepository.UpdateCustomer(customerModel);
+            }
+
+            return new Response<bool>() { ResultType = ResultType.Error, Messages = new List<string>() { Constants.CustomerNotFound } };
+        }
+
+        private async Task<bool> CheckIfCustomerExists(string email)
+        {
+            var customer = await customerRepository.GetCustomer(email);
+            if (customer.HasResult)
+                return true;
+
+            return false;
         }
     }
 }
